Always fetch tailscale apt-key on install
This changes the behavior of how the Tailscale apt repository is
configured.  It always attempts to download the latest GPG key from
Tailscale, even if the /etc/apt/sources.d/tailscale.list file exists.

This handles both the case where the key has been updated by Tailscale,
and the case where the key was lost during a UniFi upgrade (but
tailscale.list was retained), at the cost of an occasional extra https
request to pkgs.tailscale.com.